Fix hang when dragging files from file picker onto QWebEngineView
The method WebContentsAdapter::updateDragPosition actively waits for the
UpdateDragCursor message, sent by the renderer. This active wait does
not work whenever we're currently in a base::MessageLoop::RunTask call,
because of its internal recursion guard nestable_tasks_allowed. Add a
check for nestable_tasks_allowed and bail out if we know that the active
wait will fail. This fixes the hang.

Ensure that the modal file picker dialog is shown outside of
base::MessageLoop::RunTask. This enables drag 'n drop updates from the
file picker to QWebEngineView.

Task-number: QTBUG-56488
Change-Id: Ia13ada9c19e7780e12ca633ab1caeac352aca2a9
Reviewed-by: Viktor Engelmann <viktor.engelmann@qt.io>
Reviewed-by: Michael Brüning <michael.bruning@qt.io>